Fix rebase
Support passing arrays for query parameters (#10)

* Add `block_number` array query support

The `/{chain}/timestamp` endpoint can now accept a single or an array
of block numbers (comma-separated). A maximum number of values to be
parsed can be set using the MAX_ELEMENTS_QUERIED env variable (will be
shared with `/{chain}/blocknum`).

Related: #5

* Refactor API responses to serve array queries

- Changed the default Clickhouse DB response format to `JSONObjectEachRow`
in order to have query parameters matching with the response value.
- JSON responses are parsed directly from the schemas to ensure
consistency.

Closes: #5