Add CLI build with Bun and Commander
The build generates a CLI for running the Bun API. The config
has been changed to include program argument parameters. In order
to run the new `app` using custom directives (e.g port, hostname)
a new function called `generateApp()` had to be created in order for
the tests to be ran with the same app (previously `export default`).

Closes: #7