Merge develop (#15)
* Support passing arrays for query parameters (#10)

* Add `block_number` array query support

The `/{chain}/timestamp` endpoint can now accept a single or an array
of block numbers (comma-separated). A maximum number of values to be
parsed can be set using the MAX_ELEMENTS_QUERIED env variable (will be
shared with `/{chain}/blocknum`).

Related: #5

* Refactor API responses to serve array queries

- Changed the default Clickhouse DB response format to `JSONObjectEachRow`
in order to have query parameters matching with the response value.
- JSON responses are parsed directly from the schemas to ensure
consistency.

Closes: #5

* 0.1.0

* Add CLI build with Bun and Commander (#12)

* Add CLI build with Bun and Commander

The build generates a CLI for running the Bun API. The config
has been changed to include program argument parameters. In order
to run the new `app` using custom directives (e.g port, hostname)
a new function called `generateApp()` had to be created in order for
the tests to be ran with the same app (previously `export default`).

Closes: #7

* Fix build workflow

* Skip default values check in config test

* 0.2.0