Remove `chains` import for `getBlock` query
The import of the `chains` variable from `fetch/openapi.ts` caused
a DB query to be executed once the module was loaded. This behavior
was also preventing testing any function inside the `queries.ts` module.

The fix makes the `getBlock` function explicitly query the DB for
supported chains (as they might also get updated over time). The
function is mocked in the tests using the latest Bun feature (1.0.8+)
of *mock modules* in order to allow "offline" testing.